ref: move referrer_ids module out of sentry.types
